comments removed, code was cleaned
